requirements.txt
streamlit

pandas
numpy
scipy
scikit-learn
google-generativeai